requirements.txt
# DiaFootAI Requirements
# Python 3.10+ required

# ============================================
# Core Deep Learning Frameworks
# ============================================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Apple Silicon Optimization (MLX)
mlx>=0.4.0
mlx-lm>=0.0.6

# ============================================
# Computer Vision & Image Processing
# ============================================
opencv-python>=4.8.0
opencv-python-headless>=4.8.0  # For server environments
Pillow>=10.0.0
scikit-image>=0.21.0
imageio>=2.31.0

# ============================================
# Segmentation Models
# ============================================
segmentation-models-pytorch>=0.3.3
timm>=0.9.7  # PyTorch Image Models (for encoders)

# ============================================
# Data Augmentation
# ============================================
albumentations>=1.3.1
imgaug>=0.4.0

# ============================================
# Scientific Computing
# ============================================
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ============================================
# Experiment Tracking & Visualization
# ============================================
wandb>=0.15.0
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ============================================
# Configuration & CLI
# ============================================
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
typer>=0.9.0
rich>=13.5.0  # Beautiful terminal output

# ============================================
# Data Handling
# ============================================
h5py>=3.9.0
pyarrow>=13.0.0
fastparquet>=2023.8.0

# ============================================
# Model Export & Optimization
# ============================================
onnx>=1.14.0
onnxruntime>=1.15.0
coremltools>=7.0  # For iOS export

# ============================================
# Jupyter & Notebooks
# ============================================
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
tqdm>=4.66.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# ============================================
# Code Quality
# ============================================
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.3.0

# ============================================
# API & Web (for future backend)
# ============================================
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
httpx>=0.24.0

# ============================================
# Utilities
# ============================================
python-dateutil>=2.8.0
requests>=2.31.0
aiofiles>=23.2.0
tenacity>=8.2.0  # Retry logic
loguru>=0.7.0  # Better logging

# ============================================
# Medical Image Specific (Optional)
# ============================================
# pydicom>=2.4.0  # If working with DICOM
# nibabel>=5.1.0  # If working with NIfTI

# ============================================
# Mobile Export
# ============================================
# tensorflow>=2.13.0  # For TFLite export (install separately if needed)
# tf2onnx>=1.14.0
